requirements.txt
# Core
requests==2.32.3
python-dotenv==1.0.1

# Google Sheets
gspread==6.1.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1

# UI
streamlit==1.40.2
streamlit-option-menu==0.4.0
streamlit-extras==0.5.0

# Image processing
pillow==11.0.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Scheduler
apscheduler==3.10.4

# Data display
pandas==2.2.3
plotly==5.24.1

# Build
pyinstaller==6.11.1
